+++ fiddles/vue/fiddle-0010-AceEditorHwJs/src/components/ace-editor/OutlineFoldNodeInterface.ts
@@ -1,0 +1,8 @@
+import type { Ace } from 'ace-builds'
+
+export interface OutlineFoldNodeInterface {
+  range: Ace.Range
+  text: string
+  children: OutlineFoldNodeInterface[]
+  parent: OutlineFoldNodeInterface
+}