+++ CHANGELOG.md
@@ -2475,3 +2475,5 @@ Alternately, the _Daily Diary of my Programming Dreams_.
   1.  Added [fiddles/vue/fiddle-0009-SkeletonLoaderJs](fiddles/vue/fiddle-0009-SkeletonLoaderJs)
 
   
+* Added [fiddles/vue/fiddle-0010-AceEditorHw](fiddles/vue/fiddle-0010-AceEditorHw)
+* Added [fiddles/vue/fiddle-0010-AceEditorHwJs](fiddles/vue/fiddle-0010-AceEditorHwJs)